move testsuite into "default" folder
