+++ composer.json
@@ -29,8 +29,8 @@
 	},
 	"autoload-dev": {
 		"psr-4": {
-			"staabm\\PHPStanDba\\Tests\\Fixture\\": "tests/Fixture/",
-			"staabm\\PHPStanDba\\Tests\\": "tests/"
+			"staabm\\PHPStanDba\\Tests\\Fixture\\": "tests/default/Fixture/",
+			"staabm\\PHPStanDba\\Tests\\": "tests/default/"
 		}
 	},
 	"scripts": {
