+++ tests/default/DbaInferenceTest.php
@@ -57,8 +57,8 @@ class DbaInferenceTest extends TypeInferenceTestCase
     public static function getAdditionalConfigFiles(): array
     {
         return [
-            __DIR__.'/../config/stubFiles.neon',
-            __DIR__.'/../config/extensions.neon',
+            __DIR__.'/../../config/stubFiles.neon',
+            __DIR__.'/../../config/extensions.neon',
         ];
     }
 }
