+++ tests/default/Fixture/StringableObject.php
