+++ tests/default/SyntaxErrorInPreparedStatementMethodRuleTest.php
