+++ tests/default/UnresolvablePdoStatementRuleTest.php
@@ -25,7 +25,7 @@ class UnresolvablePdoStatementRuleTest extends AbstractServiceAwareRuleTestCase
 
     protected function getRule(): Rule
     {
-        return $this->getRuleFromConfig(PdoStatementExecuteMethodRule::class, __DIR__.'/../config/dba.neon');
+        return $this->getRuleFromConfig(PdoStatementExecuteMethodRule::class, __DIR__.'/../../config/dba.neon');
     }
 
     public function testSyntaxErrorInQueryRule(): void
