+++ tests/default/data/mysqli-escape.php
