+++ tests/default/data/mysqli.php
