+++ tests/default/data/pdo-column-count.php
