+++ tests/default/data/pdo-stmt-fetch.php
