+++ tests/default/data/syntax-error-in-prepared-statement.php
