+++ tests/default/data/syntax-error-in-query-function.php
