+++ tests/default/data/syntax-error-in-query-method.php
