+++ tests/default/data/unresolvable-pdo-statement.php
