+++ tests/default/data/unresolvable-query-in-function.php
