+++ tests/default/data/unresolvable-query-in-method.php
